pyproject.toml
[tool.poetry]
name = "django_pokeapi"
version = "1.0.0"
description = "Django PokeAPI"
authors = ["David Stejskal <stejski@post.cz>"]

[tool.poetry.dependencies]
python = "~3.12"  # Hard Lock

# Pydantic
pydantic = "^2.7.1"
pydantic-extra-types = "^2.7"

# Django Dependencies
django = "~5.0"  # Hard Lock
django-ninja = "~1.3.0"
django-environ = "^0.11"

# Other
httpx = {extras = ["http2"], version = "^0.27.0"}
asgiref = "^3.8.0"
psycopg = {extras = ["binary"], version = "^3.2.2"}
celery = "^5.3.4"
redis = "^5.0.1"
requests = "2.32.2"
orjson = "^3.10.18 "
psycopg2-binary = "^2.9.3"


[tool.poetry.dev-dependencies]
black = "^24.4.0"
pylint = "^3.1.0"
isort = "^5.13.2"
debugpy = "^1.8.0"
celery-types = "^0.20.0"

# Django Typing Support
django-stubs = "~5.0"   # Hard Lock - must follow django version


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.pyright]
exclude = ["**/__pycache__", ".venv", "**/migrations/**"]

[tool.ruff]

reportMissingImports = true
reportMissingTypeStubs = false
reportInvalidStubStatement = true
reportIncompleteStub = true

strictListInference = true
strictDictionaryInference = true
strictSetInference = true

reportUnknownLambdaType = true
reportUnknownParameterType = true

reportMissingParameterType = true
reportMissingTypeArgument = "warning"

reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true
reportUntypedBaseClass = true
reportUntypedNamedTuple = true

reportMatchNotExhaustive = true
reportPropertyTypeMismatch = true
reportMissingModuleSource = true
reportIncompatibleMethodOverride = true

reportDuplicateImport = true
reportShadowedImports = true
reportWildcardImportFromLibrary = true
reportPrivateImportUsage = false

reportIncompatibleVariableOverride = "warning"
reportInconsistentConstructor = "error"
reportFunctionMemberAccess = "warning"
reportConstantRedefinition = "warning"
reportOverlappingOverload = "warning"
reportPrivateUsage = "warning"
reportTypeCommentUsage = "warning"
reportUnnecessaryIsInstance = "warning"
reportUnnecessaryCast = "warning"
reportUnnecessaryContains = "warning"
reportUnnecessaryTypeIgnoreComment = "warning"

reportUnusedClass = true
reportUnusedImport = true
reportUnusedFunction = true
reportUnusedVariable = true
reportUnusedExpression = true

reportImplicitStringConcatenation = "none"
reportUnusedCallResult = "none"
